refactor: isolate Redis browser state
